Crate filedes

The top-level module filedes contains convenience / test stuff for playing with file descriptors.

The most interesting one is probably unix_socket_pair.

Modules

ring

Functions

add_tmpfile_to_ring
Uses mkstemp to create an open, unlinked temporary file and add its file descriptor to the Ring structure.

add_two_sockets_to_ring
Creates a new pair of sockets with {unix_socket_pair} and adds it to the ring, and returns the number of sockets added. This is the easiest way to get throwaway file descriptor outside wrapping libc::mkstemp (see add_tmpfile_to_ring for that) (:

connect_to_socket

make_socket_addr

server_socket

setup

teardown

unix_socket_pair
Creates a socketpair in the UNIX domain and returns it.