Struct filedes::ring::Ring

pub struct Ring {
    pub count: u64,
    // some fields omitted
}

A ring buffer containing file descriptors.

You can stuff FDs in with the add method, and iterate over them one by one using the iterator structure returned by iter.

Fields

count

The number of file descriptors contained in the ring buffer.

Methods

impl<'a> Ring

fn add(&mut self, thing: &StashableThing) -> Result<()>

Adds an FD to a Ring, updating the count of contained FDs. Closing the FD to free up resources is left to the caller.

Errors

Bad(nix::Error) - if any unforeseen condition occurs

Limit(nix::Error) - if the socket would block or any other limit runs over.

fn pop(&mut self) -> Result<StashedThing>

Removes and returns the head of the fd ring, updating count.

fn iter(&self) -> RingIter

Returns an iterator on the FDs contained in the ring buffer